[202411] fix dhcp_relay issue on marvell-teralynx (#3638)
Fixes #3637

Issue is applicable only to 202411 branch.

PR #3424 altered the platform string for Marvell Prestera to "marvell" causing it to become a substring of marvell-teralynx.

marvell-teralynx platform supports DHCP trap priority, which is crucial for dhcp packets getting trapped to the CPU. However, Marvell Prestera platform does not support trap priority.

Due to a substring match issue, the trap priority is not being applied correctly on the teralynx platform.
This results in DHCP packets not being trapped to the CPU.